Thanks Kev
I don’t think Column O should be populated with disbursement figures? If you look at other matters where there is WIP on the file, Column O only includes WIP and not disbursements as these are dealt with separately. Can we update the rule so that Column O is not populated when there are only disbursements on the file?

Many thanks

Caley 

Caley Collins 
Principal Associate  
Weightmans LLP


git-tfs-id: [http://svr-liv-tfs-02:8080/tfs/DefaultCollection]$/Reporting/Report Structure 2016;C18541